added basic RPN evaluation, not tested

+++ CMPM121Framework-main/Assets/Scripts/Levels/Enemy.cs
@@ -4,5 +4,5 @@ public class Enemy
     public int sprite;
     public int hp;
     public int speed;
-    public int damage;
+    public float damage; // switch to int?
 }
